+++ CHANGELOG.md
@@ -1,10 +1,19 @@
 # Release Notes
 
-## [Unreleased](https://github.com/laravel/nexmo-notification-channel/compare/v2.1.0...master)
+## [Unreleased](https://github.com/laravel/nexmo-notification-channel/compare/v2.2.0...master)
 
 ### Changed
 - Dropped support for Laravel 5.8 ([78bc3f9](https://github.com/laravel/nexmo-notification-channel/commit/78bc3f92091f7cd38cdb27de1df845d12f263f24))
 - Dropped support for PHP 7.1 ([858f0cb](https://github.com/laravel/nexmo-notification-channel/commit/858f0cb55c5a3bea671c10f7737926c8c8ffee2c))
+
+
+## [v2.2.0 (2019-10-01)](https://github.com/laravel/nexmo-notification-channel/compare/v2.1.0...v2.2.0)
+
+### Added
+- Implement shortcode support ([#24](https://github.com/laravel/nexmo-notification-channel/pull/24))
+
+### Changed
+- Upgrade to nexmo/laravel 2.0 ([#22](https://github.com/laravel/nexmo-notification-channel/pull/22))
 
 
 ## [v2.1.0 (2019-07-30)](https://github.com/laravel/nexmo-notification-channel/compare/v2.0.0...v2.1.0)